requirements.txt
streamlit
python-dotenv
pandas
matplotlib
faiss-cpu
# Eliminamos transformers, sentence-transformers y torch
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-text-splitters==0.0.1
langchain-huggingface
huggingface_hub
duckduckgo-search